+++ dart/lib/src/enricher/io_enricher_event_processor.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'dart:io';
+import 'dart:isolate';
 
 import '../event_processor.dart';
 import '../protocol.dart';
@@ -63,6 +64,7 @@ class IoEnricherEventProcessor extends EventProcessor {
   Map<String, dynamic> _getDartContext(bool includePii) {
     final args = Platform.executableArguments;
     final packageConfig = Platform.packageConfig;
+    final isolate = Isolate.current.debugName;
 
     String? executable;
     if (includePii) {
@@ -83,6 +85,7 @@ class IoEnricherEventProcessor extends EventProcessor {
     return <String, dynamic>{
       if (packageConfig != null) 'package_config': packageConfig,
       'number_of_processors': Platform.numberOfProcessors,
+      if (isolate != null) 'isolate': isolate,
       // The following information could potentially contain PII
       if (includePii) ...{
         'executable': executable,