+++ dart/lib/src/http_client/failed_request_client.dart
@@ -143,10 +143,7 @@ class FailedRequestClient extends BaseClient {
   }
 
   @override
-  void close() {
-    // See https://github.com/getsentry/sentry-dart/pull/226#discussion_r536984785
-    _client.close();
-  }
+  void close() => _client.close();
 
   // See https://develop.sentry.dev/sdk/event-payloads/request/
   Future<void> _captureEvent({