+++ dart/lib/src/hub.dart
@@ -400,6 +400,11 @@ class Hub {
         SentryLevel.warning,
         "Instance is disabled and this 'getSpan' call is a no-op.",
       );
+    } else if (!_options.isTracingEnabled()) {
+      _options.logger(
+        SentryLevel.info,
+        "Tracing is disabled and this 'getSpan' returns null.",
+      );
     } else {
       final item = _peek();
 
@@ -417,6 +422,11 @@ class Hub {
       _options.logger(
         SentryLevel.warning,
         "Instance is disabled and this 'captureTransaction' call is a no-op.",
+      );
+    } else if (!_options.isTracingEnabled()) {
+      _options.logger(
+        SentryLevel.info,
+        "Tracing is disabled and this 'captureTransaction' call is a no-op.",
       );
     } else if (!transaction.finished) {
       _options.logger(