+++ dart/lib/src/tracing.dart
@@ -3,3 +3,4 @@ export 'sentry_sampling_context.dart';
 export 'sentry_span_context.dart';
 export 'sentry_span_interface.dart';
 export 'noop_sentry_span.dart';
+export 'invalid_sentry_trace_header_exception.dart';