+++ flutter/test/sentry_navigator_observer_test.dart
@@ -3,9 +3,16 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 
+import 'mocks.dart';
 import 'mocks.mocks.dart';
 
 void main() {
+  late Fixture fixture;
+
+  setUp(() {
+    fixture = Fixture();
+  });
+
   group('RouteObserverBreadcrumb', () {
     test('happy path with string route agrument', () {
       const fromRouteSettings = RouteSettings(
@@ -140,12 +147,12 @@ void main() {
           settings: settings,
         );
 
-    RouteSettings routeSettings(String name, [Object? arguments]) =>
+    RouteSettings routeSettings(String? name, [Object? arguments]) =>
         RouteSettings(name: name, arguments: arguments);
 
     test('Test recording of Breadcrumbs', () {
       final hub = MockHub();
-      final observer = SentryNavigatorObserver(hub: hub);
+      final observer = fixture.getSut(hub: hub);
 
       final to = routeSettings('to', 'foobar');
       final previous = routeSettings('previous', 'foobar');
@@ -166,7 +173,7 @@ void main() {
 
     test('No arguments', () {
       final hub = MockHub();
-      final observer = SentryNavigatorObserver(hub: hub);
+      final observer = fixture.getSut(hub: hub);
 
       final to = routeSettings('to');
       final previous = routeSettings('previous');
@@ -187,7 +194,7 @@ void main() {
 
     test('No arguments & no name', () {
       final hub = MockHub();
-      final observer = SentryNavigatorObserver(hub: hub);
+      final observer = fixture.getSut(hub: hub);
 
       final to = route(null);
       final previous = route(null);
@@ -210,7 +217,7 @@ void main() {
           );
 
       final hub = MockHub();
-      final observer = SentryNavigatorObserver(hub: hub);
+      final observer = fixture.getSut(hub: hub);
 
       final to = route();
       final previous = route();
@@ -226,5 +233,79 @@ void main() {
         ).data,
       );
     });
+
+    test('route name as transaction', () {
+      final hub = _MockHub();
+      final observer = fixture.getSut(
+        hub: hub,
+        setRouteNameAsTransaction: true,
+      );
+
+      final to = routeSettings('to');
+      final previous = routeSettings('previous');
+
+      observer.didPush(route(to), route(previous));
+      expect(hub.scope.transaction, 'to');
+
+      observer.didPop(route(to), route(previous));
+      expect(hub.scope.transaction, 'previous');
+
+      observer.didReplace(newRoute: route(to), oldRoute: route(previous));
+      expect(hub.scope.transaction, 'to');
+    });
+
+    test('route name does nothing if null', () {
+      final hub = _MockHub();
+      final observer = fixture.getSut(
+        hub: hub,
+        setRouteNameAsTransaction: true,
+      );
+
+      hub.scope.transaction = 'foo bar';
+
+      final to = routeSettings(null);
+      final previous = routeSettings(null);
+
+      observer.didPush(route(to), route(previous));
+      expect(hub.scope.transaction, 'foo bar');
+    });
+
+    test('disabled route as transaction', () {
+      final hub = _MockHub();
+      final observer =
+          fixture.getSut(hub: hub, setRouteNameAsTransaction: false);
+
+      final to = routeSettings('to');
+      final previous = routeSettings('previous');
+
+      observer.didPush(route(to), route(previous));
+      expect(hub.scope.transaction, null);
+
+      observer.didPop(route(to), route(previous));
+      expect(hub.scope.transaction, null);
+
+      observer.didReplace(newRoute: route(to), oldRoute: route(previous));
+      expect(hub.scope.transaction, null);
+    });
   });
+}
+
+class Fixture {
+  SentryNavigatorObserver getSut({
+    required Hub hub,
+    bool setRouteNameAsTransaction = false,
+  }) {
+    return SentryNavigatorObserver(
+      hub: hub,
+      setRouteNameAsTransaction: setRouteNameAsTransaction,
+    );
+  }
+}
+
+class _MockHub extends MockHub {
+  final Scope scope = Scope(SentryOptions(dsn: fakeDsn));
+  @override
+  void configureScope(ScopeCallback? callback) {
+    callback?.call(scope);
+  }
 }